ai_pipeline/requirements.txt
pymongo
python-dotenv
pdfplumber
requests
pdf2image
pytesseract
pillow
# google-genai